pyproject.toml
[project]
name = "callnito"
version = "0.1.0"
description = "AI Call Center Scam Detector"
readme = "README.md"
requires-python = ">=3.10,<3.12" 

dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "dotenv>=0.9.9",
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "librosa>=0.11.0",
    "nemo-toolkit[asr]==2.4.0",
    "soundfile>=0.13.1",
    "streamlit>=1.51.0",
    "onnx==1.18.0",
    "protobuf==5.29.5",
    "streamlit-audiorec==0.1.3",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "line-bot-sdk>=3.0.0",
    "requests",
    "langchain-chroma>=0.2.2",
    "langchain-huggingface>=0.3.1",
    "sentence-transformers>=5.1.2",
    "chromadb>=0.6.3",
    "pandas>=2.3.3",
    "langchain-text-splitters>=0.3.11",
    "scipy",
    "scikit-learn",
    "numba",
]